requirements.txt
langchain==0.3.16
langchain_community==0.3.16
langchain_core==0.3.32
langchain_openai==0.3.2
python-dotenv==1.0.1
streamlit==1.32.0
